Update ios productname/headers to separate targets
Why:
We would like to be able to build the BT version of the app without
including any GPS location specific code. Currently there is some
location specific code in AppDelegate.m and some location specific swift
files being included in the BT build. We would like to start the process
of segmenting the application code by target.

This commit:
Moves Location specific code to the GPS group. It separates
AppDelegate.m into two AppDelegates, one for each target. We set up
swift bridging headers for each target. Additionally we restored some
defaults to the library and framework search paths and updated the
Product names to match the Target names of GPS and BTE as this is the
default that is expected by xcode build process and allows us to use the
default search paths provided by xcode.

Note that this commit changes no functionality.

Note that the BTE target is still building with GPS only pods currently
that are being defined from the react-native side and will need to be
removed in the BTE target in a future commit. Specifically RN
LocationAlways and @maurn85/react-native-background-geolocation